+++ admin/custom/assets/Components-CTP1qEfz.js
@@ -1,0 +1,2 @@
+import{t as e}from"./DataSolectrusItemsEditor-IJWvprMv.js";var t={DataSolectrusItemsEditor:e};export{t as default};
+//# sourceMappingURL=Components-CTP1qEfz.js.map